Default setup-env.sh
The setup-env.sh script is used when building
the container. When the container is built on
the production build server, it contains proxy
and mirror details needed to complete the build.

The blank implementation that is checked into
the repository in this commit is the baseline
environment used when building locally.